+++ plugin/test/leiningen/test/cljsbuild/config.clj
@@ -37,6 +37,7 @@
         :jar true
         :notify-command ["g"]
         :warn-on-undeclared false
+        :incremental false
         :compiler
           {:output-to "h"
            :output-dir "i"